+++ src/architecture/hooks/use-headless-context.ts
@@ -1,11 +1,10 @@
-import { Signal } from '@preact/signals';
 import { createContext, useContext } from 'react';
 
-export const HeadlessContext = createContext<Record<string, Signal>>(undefined!);
+export const HeadlessContext = createContext<Record<string, unknown>>(undefined!);
 
 /**
  * Allows the headless context to be retrieved in a child component.
- * @returns {Record<string, Signal<any>>} The headless context.
+ * @returns {Record<string, unknown>} The headless context.
  */
 const useHeadlessContext = () => {
   return useContext(HeadlessContext);