+++ src/components/form/radio-group/radio-item.props.ts
@@ -1,0 +1,16 @@
+import type { ChildrenType } from 'architecture/headless-interfaces.ts';
+
+export interface RadioItemRenderProps {
+  checked: boolean;
+  readOnly?: boolean;
+}
+
+export interface RadioItemElementProps {
+  htmlFor: string;
+}
+
+export interface RadioItemProps {
+  value?: unknown;
+  className?: string;
+  children?: ChildrenType<RadioItemRenderProps>;
+}