+++ CHANGELOG.md
@@ -9,7 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Added
 
-- Create a new toggle component and created a simple storybook test.
+- Created a new toggle component and created a simple storybook test.
+- Created a new radio group component and a simple storybook test.
+- Started using the new `useEvent` hook in place of `useCallback` for the event functions, this will decrease the number of times the event functions are created.
+
+## Changed
+
+- Installed the correct react signals version of the library since the previous version was causing issues.
 
 ## [0.0.9] - 2025-06-28